refactor: cardNumber 내부 로직 분리
