+++ hooks/src/lib/CardNumber/useCardNumbers.ts
@@ -2,6 +2,11 @@ import { useState } from 'react';
 
 import useCardNumbersValidate from './useCardNumbersValidate';
 import { cardTypeRules } from './cardTypeRules';
+import {
+  getCardType,
+  getFormattedCardNumber,
+  getCardNumberMaxLength,
+} from './cardNumberUtil';
 
 type CardType = keyof typeof cardTypeRules;
 
@@ -21,58 +26,7 @@ const useCardNumbers = (): CardNumbersResult => {
   const { errorMessage, validateCardNumbers, validateCardNumberBlur } =
     useCardNumbersValidate();
 
-  const cardNumberMaxLength = cardType ? cardTypeRules[cardType].length : 16;
-
-  const getCardType = (cardNumber: string) => {
-    if (cardNumber.startsWith('4')) {
-      return 'Visa';
-    }
-
-    if (cardNumber.startsWith('5')) {
-      const firstTwoDigits = parseInt(cardNumber.substring(0, 2));
-
-      if (firstTwoDigits >= 51 && firstTwoDigits <= 55) {
-        return 'Master';
-      }
-    }
-
-    if (cardNumber.startsWith('34') || cardNumber.startsWith('37')) {
-      return 'Amex';
-    }
-
-    if (cardNumber.startsWith('36')) {
-      return 'Diners';
-    }
-
-    if (cardNumber.startsWith('62')) {
-      if (cardNumber.length >= 3) {
-        const firstThreeDigits = parseInt(cardNumber.substring(0, 3));
-        if (firstThreeDigits >= 624 && firstThreeDigits <= 626) {
-          return 'UnionPay';
-        }
-      }
-
-      if (cardNumber.startsWith('622')) {
-        if (cardNumber.length >= 6) {
-          const firstSixDigits = parseInt(cardNumber.substring(0, 6));
-          if (firstSixDigits >= 622126 && firstSixDigits <= 622925) {
-            return 'UnionPay';
-          }
-        }
-      }
-
-      if (cardNumber.startsWith('628')) {
-        if (cardNumber.length >= 4) {
-          const firstFourDigits = parseInt(cardNumber.substring(0, 4));
-          if (firstFourDigits >= 6282 && firstFourDigits <= 6288) {
-            return 'UnionPay';
-          }
-        }
-      }
-    }
-
-    return null;
-  };
+  const cardNumberMaxLength = getCardNumberMaxLength(cardType);
 
   const handleCardNumberChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setCardType(getCardType(e.target.value));
@@ -85,31 +39,7 @@ const useCardNumbers = (): CardNumbersResult => {
     validateCardNumberBlur(cardNumbers, cardType);
   };
 
-  const getFormattedCardNumber = (cardNumbers: string) => {
-    if (!cardType) return cardNumbers;
-
-    const formattedCardNumbers = cardNumbers.replace(/\s/g, '');
-    const format = cardTypeRules[cardType].format;
-    let result = '';
-    let currentIndex = 0;
-
-    format.forEach((length) => {
-      const segment = formattedCardNumbers.slice(
-        currentIndex,
-        currentIndex + length
-      );
-      if (segment) {
-        result +=
-          segment +
-          (currentIndex + length < formattedCardNumbers.length ? '-' : '');
-      }
-      currentIndex += length;
-    });
-
-    return result;
-  };
-
-  const formattedCardNumbers = getFormattedCardNumber(cardNumbers);
+  const formattedCardNumbers = getFormattedCardNumber(cardNumbers, cardType);
 
   return {
     cardNumbers,
